Fix Linux/OSX compatibility
All file replacements must be lowercase in the actual filesystem for
Linux/OSX compatibility.

On the other hand, original files should be UpperCase formatted to
minimize inconsistencies in the include lua function between the Windows
release and the Aspyr port.

See here for a full writeup and explanation of the underlying quirks:
https://github.com/civfanatics/CQUI_Community-Edition/issues/116#issuecomment-667436886